fix bugs for sing battle game

+++ Android/scenes/ktv_singbattle/src/main/java/io/agora/scene/ktv/singbattle/widget/rankList/RankItem.java
@@ -5,5 +5,5 @@ public class RankItem {
     public String userName;
     public String poster;
     public int songNum;
-    public double score;
+    public int score;
 }
